+++ docs/changelog.md
@@ -15,6 +15,7 @@ This will probably be the last update for a while until I figure out raster and 
 - Disallowed setting `layer.size`: will re-enable individually for raster, vector, etc as convenience wrappers for transformation.
 - Added `layer.x`, `layer.y` convenience properties.
 - Tweaked `layer.position` to be relative to the top left, not bottom left.
+- Added a binding class to `layer.styles` such that `.append` (etc) work.
 - The beginnings of a `layer.adjusts` property, a library of tweaks and dials that allows, say, `layer.adjusts.white_balance.tint = -0.4`. (Not completed!)
 
 ### 0.0.4 (2020-01-15)